# cms-api/src/interfaces/http/openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KAPP CMS API",
    "version": "v1"
  },
  "servers": [
    {
      "description": "rest-temporary",
      "url": "https://qp3fgsp42g.execute-api.ap-southeast-1.amazonaws.com/rest/api"
    },
    {
      "description": "dev",
      "url": "https://sztg7uncla.execute-api.ap-southeast-1.amazonaws.com/dev/api"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "operationId": "ListPosts",
        "tags": [ "Posts" ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "filter post status",
            "required": false,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "example": [{
                    "status": "for-approval"
                  },
                  {
                    "status": "for-revision"
                  },
                  {
                    "status": "published"
                  },
                  {
                    "status": "scheduled"
                  },
                  {
                    "status": "draft"
                  }]
                }
              }
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "filter location",
            "required": false,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "example": [{
                    "location": "Quezon City"
                  }]
                }
              }
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "filter post based filtered tab",
            "required": false,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "example": [{
                    "date": "2019-10-31 03:00:57"
                  }]
                }
              }
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "filter post category",
            "required": false,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "example": [{
                    "category": "Jobs"
                  }]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "500": {
            "description": "List of all posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createInitialPost",
        "tags": [ "Posts" ],
        "requestBody": {
          "description": "Creates an initial post",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostInitial"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Initial post has started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostInitial"
                }
              }
            }
          },
          "400": {
            "description": "Not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "showPost",
        "tags": [ "Posts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of post to show",
            "required": true,
            "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
          }
        ],
        "responses": {
          "200": {
            "description": "Return post with given id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "savePost",
        "tags": [ "Posts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of post to update",
            "required": true,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Post Draft new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Post Draft updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}/save": {
      "post": {
        "operationId": "SaveRawPost",
        "tags": [ "Posts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of raw post",
            "required": true,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Post Raw new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Post Raw updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}/revise": {
      "post": {
        "operationId": "RevisePost",
        "tags": [ "Posts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of post to revise",
            "required": true,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Post Draft new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Post Draft updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}/publish": {
      "post": {
        "operationId": "PublishPost",
        "tags": [ "Posts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of post to publish",
            "required": true,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Post Draft new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Post Draft updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/advisories": {
      "get": {
        "operationId": "ListAdvisories",
        "tags": [ "Advisories" ],
        "parameters": [
          {
            "name": "draft",
            "in": "query",
            "description": "filter advisory draft",
            "required": false,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Advisory"
                  },
                  "example": [{
                    "draft": true
                  }]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all advisory",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Advisory"
                  }
                }
              }
            }
          },
          "500": {
            "description": "List of all advisories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAdvisoryDraft",
        "tags": [ "Advisories" ],
        "requestBody": {
          "description": "Creates a draft of a advisory",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisoryDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Advisory draft has started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvisoryDraft"
                }
              }
            }
          },
          "400": {
            "description": "Not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/advisories/{id}": {
      "get": {
        "operationId": "showAdvisory",
        "tags": [ "Advisories" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of advisory to show",
            "required": true,
            "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Advisory"
                  }
                }
          }
        ],
        "responses": {
          "200": {
            "description": "Return advisory with given id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Advisory"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "saveAdvisory",
        "tags": [ "Advisories" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of advisory to update",
            "required": true,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Advisory"
                  }
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Advisory Draft new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Advisory"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Advisoru Draft updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Advisory"
                }
              }
            }
          },
          "404": {
            "description": "Advisory not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/advisories/{id}/publish": {
      "post": {
        "operationId": "PublishAdvisory",
        "tags": [ "Advisories" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of advisory to publish",
            "required": true,
              "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Advisory"
                  }
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Advisory Draft new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Advisory"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Advisory Draft updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Advisory"
                }
              }
            }
          },
          "404": {
            "description": "Advisory not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique Identifier representing a user",
            "x-primary-key": true
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },   
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique Identifier representing a post",
            "x-primary-key": true
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "description": "Identifier representing an author"
          },
          "categoryId": {
            "type": "string"
          },
          "subCategory": {
            "type": "string"
          },
          "postId": {  
             "type": "string",
            "format": "int32",
            "description": "Unique Identifier representing a post",
            "x-primary-key": true
          },
          "title": {
            "type": "string"
          },  
          "content": {
            "type": "string"
          },  
           "priorityLevel": {
            "type": "string"
          },"source": {
            "type": "string"
          },"locationAddress": {
            "type": "string"
          },"locationDetails": {
            "type": "object"
          },"tagsOriginal": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },"tagsRetained": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },"tagsAdded": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },"tagsRemoved": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },"advisories": {
            "type": "object"
          },"attachments": {
            "type": "object"
          },"assignedUserId": {
            "type": "integer"
          },"status": {
            "type": "string"
          },"scheduledAt": {
            "type": "string"
          },"expiredAt": {
            "type": "string"
          },"publishedAt": {
            "type": "string"
          }, "user": {
            "type": "object"
          }
        }
      },
      "Advisory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique Identifier representing a advisory",
            "x-primary-key": true
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "description": "Identifier representing an author"
          },
          "categoryId": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "advisoryId": {  
             "type": "string",
            "format": "int32",
            "description": "Unique Identifier representing a advisory",
            "x-primary-key": true
          },
          "title": {
            "type": "string"
          },  
          "content": {
            "type": "string"
          },"source": {
            "type": "string"
          },"locationAddress": {
            "type": "string"
          },"locationDetails": {
            "type": "object"
          },"attachments": {
            "type": "object"
          },"draft": {
            "type": "boolean"
          },"verified": {
            "type": "boolean"
          },"publishedAt": {
            "type": "string"
          }
        }
      },
      "PostInitial": {
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique Identifier representing a draft of post",
            "x-primary-key": true,
            "default": 1
          }
        }
      },
      "AdvisoryDraft": {
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique Identifier representing a draft of advisory",
            "x-primary-key": true,
            "default": 1
          }
        }
      },
    "ValidationError": {
      "properties": {
        "type": {
          "type": "string",
          "enum": [ "ValidationError" ]
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/ValidationErrorDetail"
          }
        }
      }
    },
    "ValidationErrorDetail": {
      "properties": {
        "message": {
          "type": "string"
        },
        "path": {
          "type": "string"
        }
      }
    },
    "NotFoundError": {
      "properties": {
        "type": {
          "type": "string",
          "enum": [ "NotFoundError" ]
        },
        "message": {
          "type": "string",
          "enum": [ "Not found message" ]
        }
      }
    }, 
    "IntenalServerError": {
      "properties": {
        "type": {
          "type": "string",
          "enum": [ "NotFoundError" ]
        },
        "message": {
          "type": "string",
          "enum": [ "Not found message" ]
        }
      }
    }
    }
  }
}